requirements.txt
tensorflow-gpu==1.15
numpy
setuptools
six
Pillow
protobuf
# requests ~> 2.20.0 # for downloading
pandas
# and opencv => needs to compiled
